Move InternalKSPException to errors package in API module
This change is made to avoid classpath collisions/shadowing of
InternalKSPException.
