+++ kotlin-analysis-api/src/main/kotlin/com/google/devtools/ksp/impl/ResolverAAImpl.kt
@@ -18,17 +18,97 @@
 
 package com.google.devtools.ksp.impl
 
-import com.google.devtools.ksp.*
+import com.google.devtools.ksp.InternalKSPException
 import com.google.devtools.ksp.InternalKSPException.Companion.render
-import com.google.devtools.ksp.common.*
-import com.google.devtools.ksp.common.impl.*
+import com.google.devtools.ksp.KspExperimental
+import com.google.devtools.ksp.closestClassDeclaration
+import com.google.devtools.ksp.common.JVM_DEFAULT_ANNOTATION_FQN
+import com.google.devtools.ksp.common.JVM_DEFAULT_WITHOUT_COMPATIBILITY_ANNOTATION_FQN
+import com.google.devtools.ksp.common.JVM_STATIC_ANNOTATION_FQN
+import com.google.devtools.ksp.common.JVM_STRICTFP_ANNOTATION_FQN
+import com.google.devtools.ksp.common.JVM_SYNCHRONIZED_ANNOTATION_FQN
+import com.google.devtools.ksp.common.JVM_TRANSIENT_ANNOTATION_FQN
+import com.google.devtools.ksp.common.JVM_VOLATILE_ANNOTATION_FQN
+import com.google.devtools.ksp.common.extractThrowsAnnotation
+import com.google.devtools.ksp.common.impl.AnnotationResolutionStrategy
+import com.google.devtools.ksp.common.impl.KSNameImpl
+import com.google.devtools.ksp.common.impl.KSTypeReferenceSyntheticImpl
+import com.google.devtools.ksp.common.impl.RefPosition
+import com.google.devtools.ksp.common.impl.findOuterMostRef
+import com.google.devtools.ksp.common.impl.findRefPosition
+import com.google.devtools.ksp.common.impl.isReturnTypeOfAnnotationMethod
+import com.google.devtools.ksp.common.javaModifiers
+import com.google.devtools.ksp.common.lazyMemoizedSequence
+import com.google.devtools.ksp.getClassDeclarationByName
+import com.google.devtools.ksp.getDeclaredFunctions
+import com.google.devtools.ksp.getDeclaredProperties
 import com.google.devtools.ksp.impl.symbol.java.KSAnnotationJavaImpl
-import com.google.devtools.ksp.impl.symbol.kotlin.*
-import com.google.devtools.ksp.impl.symbol.util.*
+import com.google.devtools.ksp.impl.symbol.kotlin.AbstractKSDeclarationImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSClassDeclarationEnumEntryImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSClassDeclarationImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSErrorType
+import com.google.devtools.ksp.impl.symbol.kotlin.KSFileImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSFileJavaImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSFunctionDeclarationImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSFunctionErrorImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSFunctionImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSPropertyAccessorImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSPropertyDeclarationImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSPropertyDeclarationJavaImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSTypeAliasImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSTypeArgumentLiteImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.KSTypeImpl
+import com.google.devtools.ksp.impl.symbol.kotlin.analyze
+import com.google.devtools.ksp.impl.symbol.kotlin.explictJvmName
+import com.google.devtools.ksp.impl.symbol.kotlin.fillInDeepSubstitutor
+import com.google.devtools.ksp.impl.symbol.kotlin.findParentOfType
+import com.google.devtools.ksp.impl.symbol.kotlin.fullyExpand
+import com.google.devtools.ksp.impl.symbol.kotlin.inlineSuffix
+import com.google.devtools.ksp.impl.symbol.kotlin.internalSuffix
+import com.google.devtools.ksp.impl.symbol.kotlin.toKSDeclaration
+import com.google.devtools.ksp.impl.symbol.kotlin.toKSName
+import com.google.devtools.ksp.impl.symbol.kotlin.toKtClassSymbol
+import com.google.devtools.ksp.impl.symbol.kotlin.toLocation
+import com.google.devtools.ksp.impl.symbol.kotlin.toWildcard
+import com.google.devtools.ksp.impl.symbol.kotlin.typeArguments
+import com.google.devtools.ksp.impl.symbol.kotlin.updateFromParents
+import com.google.devtools.ksp.impl.symbol.util.BinaryClassInfoCache
 import com.google.devtools.ksp.impl.symbol.util.DeclarationOrdering
+import com.google.devtools.ksp.impl.symbol.util.extractThrowsFromClassFile
+import com.google.devtools.ksp.impl.symbol.util.getFileContent
+import com.google.devtools.ksp.impl.symbol.util.getVirtualFile
+import com.google.devtools.ksp.impl.symbol.util.hasAnnotation
+import com.google.devtools.ksp.isAbstract
+import com.google.devtools.ksp.isConstructor
+import com.google.devtools.ksp.isOpen
+import com.google.devtools.ksp.isPrivate
+import com.google.devtools.ksp.isProtected
+import com.google.devtools.ksp.isPublic
+import com.google.devtools.ksp.isVisibleFrom
 import com.google.devtools.ksp.processing.KSBuiltIns
 import com.google.devtools.ksp.processing.Resolver
-import com.google.devtools.ksp.symbol.*
+import com.google.devtools.ksp.symbol.ClassKind
+import com.google.devtools.ksp.symbol.KSAnnotated
+import com.google.devtools.ksp.symbol.KSAnnotation
+import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSClassifierReference
+import com.google.devtools.ksp.symbol.KSDeclaration
+import com.google.devtools.ksp.symbol.KSDeclarationContainer
+import com.google.devtools.ksp.symbol.KSFile
+import com.google.devtools.ksp.symbol.KSFunction
+import com.google.devtools.ksp.symbol.KSFunctionDeclaration
+import com.google.devtools.ksp.symbol.KSName
+import com.google.devtools.ksp.symbol.KSNode
+import com.google.devtools.ksp.symbol.KSPropertyAccessor
+import com.google.devtools.ksp.symbol.KSPropertyDeclaration
+import com.google.devtools.ksp.symbol.KSPropertyGetter
+import com.google.devtools.ksp.symbol.KSPropertySetter
+import com.google.devtools.ksp.symbol.KSType
+import com.google.devtools.ksp.symbol.KSTypeArgument
+import com.google.devtools.ksp.symbol.KSTypeReference
+import com.google.devtools.ksp.symbol.Modifier
+import com.google.devtools.ksp.symbol.Origin
+import com.google.devtools.ksp.symbol.Variance
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiMethod
 import com.intellij.psi.impl.file.impl.JavaFileManager
@@ -37,7 +117,14 @@ import org.jetbrains.kotlin.analysis.api.components.buildSubstitutor
 import org.jetbrains.kotlin.analysis.api.fir.symbols.KaFirSymbol
 import org.jetbrains.kotlin.analysis.api.fir.types.KaFirType
 import org.jetbrains.kotlin.analysis.api.projectStructure.KaSourceModule
-import org.jetbrains.kotlin.analysis.api.symbols.*
+import org.jetbrains.kotlin.analysis.api.symbols.KaEnumEntrySymbol
+import org.jetbrains.kotlin.analysis.api.symbols.KaFunctionSymbol
+import org.jetbrains.kotlin.analysis.api.symbols.KaJavaFieldSymbol
+import org.jetbrains.kotlin.analysis.api.symbols.KaNamedClassSymbol
+import org.jetbrains.kotlin.analysis.api.symbols.KaPropertyAccessorSymbol
+import org.jetbrains.kotlin.analysis.api.symbols.KaPropertySymbol
+import org.jetbrains.kotlin.analysis.api.symbols.KaSymbol
+import org.jetbrains.kotlin.analysis.api.symbols.KaTypeAliasSymbol
 import org.jetbrains.kotlin.analysis.api.types.KaType
 import org.jetbrains.kotlin.analysis.api.types.symbol
 import org.jetbrains.kotlin.analysis.decompiler.stub.file.ClsKotlinBinaryClassCache
